Move socket mocks to their own file
